ath10k: use raw tx
This is just a reference for anyone interested in
getting injection working with ath10k.

This just switches to raw tx mode instead of nwifi
for regular operation. Injection doesn't work yet.

I've tested this on 999.999.0.636, 10.1.467.2-1
and 10.2-00082-4. The 10.1.467.2-1 crashes as soon
as first tx completion comes in. The other two
seem to work but don't work with A-MSDU so it is
necessary to disable it:

Be sure to run:
 echo 1 64 > /sys/kernel/debug/ieee80211/phy*/ath10k/htt_max_amsdu_ampdu

Signed-off-by: Michal Kazior <michal.kazior at tieto.com>